ATR-855: moved events recognition to a separate helper and namespace
