+++ src/Acuminator/Acuminator.Analyzers/StaticAnalysis/ForbidPrivateEventHandlers/ForbidPrivateEventHandlersAnalyzer.cs
@@ -5,6 +5,7 @@ using System.Linq;
 using Acuminator.Analyzers.StaticAnalysis.PXGraph;
 using Acuminator.Utilities.DiagnosticSuppression;
 using Acuminator.Utilities.Roslyn.Semantic;
+using Acuminator.Utilities.Roslyn.Semantic.EventsRecognition;
 using Acuminator.Utilities.Roslyn.Semantic.PXGraph;
 
 using Microsoft.CodeAnalysis;
